src/requirements.txt
python-gitlab
openai
requests
flask